+++ tfjs-backend-cpu/src/kernels/Acosh.ts
@@ -17,7 +17,7 @@
 
 import {Acosh, KernelConfig} from '@tensorflow/tfjs-core';
 
-import {unaryKernelFunc} from '../utils/kernel_utils';
+import {unaryKernelFunc} from '../utils/unary_utils';
 
 export const acoshKernelFunc = unaryKernelFunc(Acosh, (xi) => Math.acosh(xi));
