+++ tfjs-backend-cpu/src/kernels/Asin.ts
@@ -17,7 +17,7 @@
 
 import {Asin, KernelConfig} from '@tensorflow/tfjs-core';
 
-import {unaryKernelFunc} from '../utils/kernel_utils';
+import {unaryKernelFunc} from '../utils/unary_utils';
 
 export const asinKernelFunc = unaryKernelFunc(Asin, (xi) => Math.asin(xi));
