+++ tfjs-backend-cpu/src/kernels/Atan.ts
@@ -17,7 +17,7 @@
 
 import {Atan, KernelConfig} from '@tensorflow/tfjs-core';
 
-import {unaryKernelFunc} from '../utils/kernel_utils';
+import {unaryKernelFunc} from '../utils/unary_utils';
 
 export const atanKernelFunc = unaryKernelFunc(Atan, (xi) => Math.atan(xi));
