+++ tfjs-backend-cpu/src/kernels/Ceil.ts
@@ -17,7 +17,7 @@
 
 import {Ceil, KernelConfig} from '@tensorflow/tfjs-core';
 
-import {unaryKernelFunc} from '../utils/kernel_utils';
+import {unaryKernelFunc} from '../utils/unary_utils';
 
 export const ceilKernelFunc = unaryKernelFunc(Ceil, (xi) => Math.ceil(xi));
