+++ tfjs-backend-cpu/src/kernels/Cos.ts
@@ -17,7 +17,7 @@
 
 import {Cos, KernelConfig} from '@tensorflow/tfjs-core';
 
-import {unaryKernelFunc} from '../utils/kernel_utils';
+import {unaryKernelFunc} from '../utils/unary_utils';
 
 export const cosKernelFunc = unaryKernelFunc(Cos, (xi) => Math.cos(xi));
