+++ tfjs-backend-cpu/src/kernels/Cosh.ts
@@ -17,7 +17,7 @@
 
 import {Cosh, KernelConfig} from '@tensorflow/tfjs-core';
 
-import {unaryKernelFunc} from '../utils/kernel_utils';
+import {unaryKernelFunc} from '../utils/unary_utils';
 
 export const coshKernelFunc = unaryKernelFunc(Cosh, (xi) => Math.cosh(xi));
