+++ tfjs-backend-cpu/src/kernels/Elu.ts
@@ -17,7 +17,7 @@
 
 import {Elu, KernelConfig} from '@tensorflow/tfjs-core';
 
-import {unaryKernelFunc} from '../utils/kernel_utils';
+import {unaryKernelFunc} from '../utils/unary_utils';
 
 export const eluKernelFunc =
     unaryKernelFunc(Elu, (xi) => xi >= 0 ? xi : (Math.exp(xi) - 1));
