+++ tfjs-backend-cpu/src/kernels/Exp.ts
@@ -17,7 +17,7 @@
 
 import {Exp, KernelConfig} from '@tensorflow/tfjs-core';
 
-import {unaryKernelFunc} from '../utils/kernel_utils';
+import {unaryKernelFunc} from '../utils/unary_utils';
 
 export const expKernelFunc = unaryKernelFunc(Exp, (xi) => Math.exp(xi));
