+++ tfjs-backend-cpu/src/kernels/Floor.ts
@@ -17,7 +17,7 @@
 
 import {Floor, KernelConfig} from '@tensorflow/tfjs-core';
 
-import {unaryKernelFunc} from '../utils/kernel_utils';
+import {unaryKernelFunc} from '../utils/unary_utils';
 
 export const floorKernelFunc = unaryKernelFunc(Floor, (xi) => Math.floor(xi));
