+++ tfjs-backend-cpu/src/kernels/IsFinite.ts
@@ -17,7 +17,7 @@
 
 import {IsFinite, KernelConfig} from '@tensorflow/tfjs-core';
 
-import {unaryKernelFunc} from '../utils/kernel_utils';
+import {unaryKernelFunc} from '../utils/unary_utils';
 
 export const isFiniteKernelFunc =
     unaryKernelFunc(IsFinite, (xi) => Number.isFinite(xi) ? 1 : 0, 'bool');
