+++ tfjs-backend-cpu/src/kernels/IsInf.ts
@@ -17,7 +17,7 @@
 
 import {IsInf, KernelConfig} from '@tensorflow/tfjs-core';
 
-import {unaryKernelFunc} from '../utils/kernel_utils';
+import {unaryKernelFunc} from '../utils/unary_utils';
 
 export const isInfKernelFunc =
     unaryKernelFunc(IsInf, (xi) => Math.abs(xi) === Infinity ? 1 : 0, 'bool');
