+++ tfjs-backend-cpu/src/kernels/LogicalNot.ts
@@ -17,7 +17,7 @@
 
 import {KernelConfig, LogicalNot} from '@tensorflow/tfjs-core';
 
-import {unaryKernelFunc} from '../utils/kernel_utils';
+import {unaryKernelFunc} from '../utils/unary_utils';
 
 export const logicalNotKernelFunc =
     unaryKernelFunc(LogicalNot, (xi) => xi ? 0 : 1, 'bool');
