+++ tfjs-backend-cpu/src/kernels/Reciprocal.ts
@@ -17,7 +17,7 @@
 
 import {KernelConfig, Reciprocal} from '@tensorflow/tfjs-core';
 
-import {unaryKernelFunc} from '../utils/kernel_utils';
+import {unaryKernelFunc} from '../utils/unary_utils';
 
 export const reciprocalKernelFunc = unaryKernelFunc(Reciprocal, (xi) => 1 / xi);
