+++ tfjs-backend-cpu/src/kernels/Round.ts
@@ -17,7 +17,7 @@
 
 import {KernelConfig, Round} from '@tensorflow/tfjs-core';
 
-import {unaryKernelFunc} from '../utils/kernel_utils';
+import {unaryKernelFunc} from '../utils/unary_utils';
 
 export const roundKernelFunc = unaryKernelFunc(Round, (xi) => {
   // The algorithm is based on banker's rounding.
