+++ tfjs-backend-cpu/src/kernels/Sin.ts
@@ -17,7 +17,7 @@
 
 import {KernelConfig, Sin} from '@tensorflow/tfjs-core';
 
-import {unaryKernelFunc} from '../utils/kernel_utils';
+import {unaryKernelFunc} from '../utils/unary_utils';
 
 export const sinKernelFunc = unaryKernelFunc(Sin, (xi) => Math.sin(xi));
