+++ tfjs-backend-cpu/src/kernels/Sinh.ts
@@ -17,7 +17,7 @@
 
 import {KernelConfig, Sinh} from '@tensorflow/tfjs-core';
 
-import {unaryKernelFunc} from '../utils/kernel_utils';
+import {unaryKernelFunc} from '../utils/unary_utils';
 
 export const sinhKernelFunc = unaryKernelFunc(Sinh, (xi) => Math.sinh(xi));
