+++ tfjs-backend-cpu/src/kernels/Sqrt.ts
@@ -17,7 +17,7 @@
 
 import {KernelConfig, Sqrt} from '@tensorflow/tfjs-core';
 
-import {unaryKernelFunc} from '../utils/kernel_utils';
+import {unaryKernelFunc} from '../utils/unary_utils';
 
 export const sqrtKernelFunc = unaryKernelFunc(Sqrt, (xi) => Math.sqrt(xi));
